requirements.txt
djangorestframework~=3.16.1
Django~=5.2.7
AgentService~=0.0.92
pydantic~=2.10.1
pydantic-settings~=2.2.1
django-browser-reload
django-cotton
django-mathfilters
django-htmx
django-tailwind
psycopg2-binary
uvicorn
whitenoise
asgiref
django_components
AgentService==0.0.141